refactor(skill-host-contracts): move AssistantEvent types into neutral package
Part of plan: skill-isolation.md (PR 2 of 34)
