+++ assistant/package.json
@@ -41,6 +41,7 @@
     "@vellumai/ces-contracts": "file:../packages/ces-contracts",
     "@vellumai/credential-storage": "file:../packages/credential-storage",
     "@vellumai/egress-proxy": "file:../packages/egress-proxy",
+    "@vellumai/skill-host-contracts": "file:../packages/skill-host-contracts",
     "archiver": "7.0.1",
     "commander": "13.1.0",
     "croner": "10.0.1",
@@ -70,7 +71,8 @@
   "bundledDependencies": [
     "@vellumai/ces-contracts",
     "@vellumai/credential-storage",
-    "@vellumai/egress-proxy"
+    "@vellumai/egress-proxy",
+    "@vellumai/skill-host-contracts"
   ],
   "overrides": {
     "lodash": "4.18.1",
